[#166665367] add test about fixWalletPspTagsValues patch function
