+++ ts/api/pagopa.ts
@@ -57,6 +57,7 @@ import {
   updateWalletUsingPUTDecoder,
   UpdateWalletUsingPUTT
 } from "../../definitions/pagopa/requestTypes";
+import { fixWalletPspTagsValues } from "../utils/wallet";
 /**
  * A decoder that ignores the content of the payload and only decodes the status
  */
@@ -133,34 +134,6 @@ type GetWalletsUsingGETExtraT = MapResponseType<
   200,
   WalletListResponse
 >;
-
-/**
- * it sanitizes psp tags avoiding no string value and string duplicates
- * @param w wallet object
- */
-const fixWalletPspTagsValues = (w: unknown) => {
-  const decoder = t.interface({
-    psp: t.interface({
-      tags: t.readonlyArray(t.unknown)
-    })
-  });
-  const decoded = decoder.decode(w);
-  if (decoded.isLeft()) {
-    return w;
-  }
-  const psp = decoded.value.psp;
-  const tags = decoded.value.psp.tags;
-  return {
-    ...decoded.value,
-    psp: {
-      ...psp,
-      tags: tags.filter(
-        (item: any, idx: number) =>
-          typeof item === "string" && tags.indexOf(item) === idx
-      )
-    }
-  };
-};
 
 /**
  *
